feat: add -A, to include all endpoints
The idea is that random_mullvad_endpoints can now randomly select
entries without replacement until the entire list has been cycled
through, then return it. The advantage of allowing this over something
like `sort -R` is that the weights are accounted for, so low weights
will naturally end up toward the bottom of the list most of the time.